pedal and time display in level selection

+++ Assets/Scripts/GameCycle/LevelAchievements.cs
@@ -2,7 +2,7 @@
 {
     public class LevelAchievements
     {
-        public float PlayerPerformanceTime;
+        public int PlayerPerformanceTime;
         public float TotalScore;
     }
 }
